requirements.txt
build==0.5.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gspread==3.7.0
httplib2==0.19.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.0
pep517==0.10.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
research-thyroid-digitake @ file:///Users/song/Dev/thyroid
rsa==4.7.2
six==1.16.0
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
